[SPARK-28996][SQL] Add tests regarding username of HiveClient
